Act trans delete BUT is completed
